+++ packages/fxa-settings/src/pages/ResetPassword/AccountRecoveryConfirmKey/en.ftl
