+++ packages/fxa-settings/src/pages/ResetPassword/AccountRecoveryResetPassword/en.ftl
