+++ packages/fxa-settings/src/pages/ResetPassword/ResetPasswordWithRecoveryKeyConfirmed/en.ftl
